Fix hsql jdbc driver deps (#5168)
Avoid non-included classes in jdbc driver dependencies.
